requirements.txt
# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.11.0

# Telegram Bot
python-telegram-bot==20.7

# Environment variables
python-dotenv==1.0.0

# Date/Time handling
pytz==2024.1

# CLI interface
click==8.1.7

